Platform.sh works with PlatformDeployer in separate module.
